Added deletion security

+++ delete_file.php
@@ -4,16 +4,22 @@ if (isset($_GET['filename'])) {
     $directory = './files/';
     $filepath = $directory . $filename;
 
+    $namewithoutext = pathinfo($filename)['filename'];
+
     if (file_exists($filepath)) {
-        if (unlink($filepath)) {
-            echo "File $filename deleted successfully.";
+        if (substr($namewithoutext, -3) !== "​") {
+            if (unlink($filepath)) {
+                echo "File $filename deleted successfully.";
+            } else {
+                echo "Error deleting file $filename.";
+            }
         } else {
-            echo "Error deleting file $filename.";
+            echo "Illegal download.";
         }
     } else {
         echo "File $filename does not exist.";
     }
 } else {
-    echo "Filename parameter is missing.";
+    echo "Filename parameter is missing or empty.";
 }
 ?>
